move CurrencyField to counter
