+++ counter/views/invoice.py
@@ -19,7 +19,7 @@ from django.db.models import F
 from django.utils import timezone
 from django.views.generic import TemplateView
 
-from accounting.models import CurrencyField
+from counter.fields import CurrencyField
 from counter.models import Refilling, Selling
 from counter.views.mixins import CounterAdminMixin, CounterAdminTabsMixin
